remapped controls
controls for both gamepad 1 and 2,
made manualdrive simpler
