+++ FtcRobotController/ManualDriveOpMode.java
@@ -11,8 +11,8 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
  * becomes relative to the field as opposed to the robot. You can
  * reset the forward heading by pressing "x".
  */
-@TeleOp(name = "Manual Drive Red")
-public class ManualDriveOpModeRedWall extends LinearOpMode {
+@TeleOp(name = "Manual Drive Blue")
+public class ManualDriveOpMode extends LinearOpMode {
     private servoDropBot robot = new servoDropBot(this);
     private boolean arcadeMode = false;
     private int gyroCalibratedCount = 0;
@@ -25,7 +25,7 @@ public class ManualDriveOpModeRedWall extends LinearOpMode {
         waitForStart();
         while (opModeIsActive()) {
 
-            robot.driveByHandRedWall(gamepad1.left_stick_y, gamepad1.left_stick_x, gamepad1.right_stick_x);
+            robot.driveByHand(gamepad1.left_stick_y, gamepad1.left_stick_x, gamepad1.right_stick_x);
             robot.scoopSetPosition(gamepad1.b, gamepad1.y, gamepad1.x);
             robot.scoopFreeRun(gamepad1.right_bumper, gamepad1.left_bumper);
             robot.manualFoundationDrop(gamepad1.dpad_down);
